% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arroz}
\alias{arroz}
\title{Data of Grain Yield of Rice}
\format{
A matrix with 20 rows and 20 columns, where each cell represents the rice
yield in grams of a 1m² subplot.
}
\source{
\doi{10.22458/urj.v11i3.2653}
}
\usage{
data("arroz")
}
\description{
This dataset contains yield data from a uniformity trial with rice
(Oryza sativa) of the cultivar Palmar 18. The trial was conducted at
Hacienda Mojica, located in the cantón of Bagaces, Guanacaste, Costa Rica.
The experiment was carried out on a 20m x 20m (400m²) plot, which was
divided into 400 basic experimental units of 1m² each. The yield of
each unit is recorded in grams.
}
\details{
\itemize{
  \item \strong{Crop:} Rice (Oryza sativa)
  \item \strong{Cultivar:} Palmar 18
  \item \strong{Planting Method:} Direct seeding with rows separated by 17.6cm.
  \item \strong{Total Area:} 20m x 20m
  \item \strong{Basic Unit:} 1m x 1m
  \item \strong{Data:} Yield in grams per basic unit.
  \item \strong{Dimensions:} The matrix has 20 rows (length, Y-axis) and 20 columns (width,
    X-axis).
}
}
\examples{
library("fielddesign")
data("arroz")
# View the first 6 rows and 6 columns.
arroz[1:6, 1:6]

}
\references{
Vargas-Rojas, J. C., & Navarro-Flores, J. R. (2019). Tamaño y forma de unidad
experimental para ensayos de rendimiento de arroz (Oryza sativa), en
Guanacaste, Costa Rica. UNED Research Journal, 11(3).
\doi{10.22458/urj.v11i3.2653}
}
\keyword{datasets}
