% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_of_variation.R
\name{coefficient_of_variation}
\alias{coefficient_of_variation}
\title{Calculate the Coefficient of Variation}
\usage{
coefficient_of_variation(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar representing the coefficient of variation.
}
\description{
This function computes the coefficient of variation (CV) for a numeric vector.
The CV is a measure of relative variability, expressed as the ratio of the standard deviation to
the mean.
}
\keyword{internal}
