% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_sum_matrix.R
\name{cumulative_sum_matrix}
\alias{cumulative_sum_matrix}
\title{Calculate 2D Cumulative Sum of a Matrix}
\usage{
cumulative_sum_matrix(x)
}
\arguments{
\item{x}{A numeric matrix with all finite numbers. The matrix must have at least 2 rows.
If the matrix has just 1 col, the result will be transposed (apply `t` after calculations).}
}
\value{
A numeric matrix representing the two-dimensional cumulative sum of `M`.
}
\description{
This function calculates the two-dimensional cumulative sum of a numeric matrix.
It first applies cumulative sum to each column, and then to each row of the resulting matrix.
}
\keyword{internal}
