% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fielddesign_app.R
\name{fielddesign_app}
\alias{fielddesign_app}
\title{Launch the Field Design Shiny Gadget}
\usage{
fielddesign_app(x = NULL)
}
\arguments{
\item{x}{An optional argument for initial yield data. Can be:
- `NULL` (default): The app will start with no data loaded, and the user can upload a file.
- A character string: Path to a file containing yield data. The file is read as a table and
  converted to a matrix.
- A matrix or data frame with the data.}
}
\value{
No return value. Called for its side effects to launch the Field Design Shiny Gadget.
}
\description{
This function launches a Shiny gadget for interactive analysis of experimental unit size and
spatial variation in yield data.
}
\examples{
if (interactive()) {
  fielddesign_app()
  # Specify the input data.
  fielddesign_app(matrix(rnorm(130, 500, 60), nrow = 10))
}

}
