% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maiz}
\alias{maiz}
\title{Data of Grain Yield of Corn}
\format{
A matrix with 20 rows and 20 columns, where each cell represents the corn
yield in grams of a 1m² subplot.
}
\source{
\doi{10.15517/isucr.v21i43.41972}
}
\usage{
data("maiz")
}
\description{
This dataset contains the yield data from a uniformity trial with corn
(Zea mays L.) of the hybrid HS5G. The trial was conducted at the Finca
Experimental de Santa Cruz, owned by the University of Costa Rica, located
in the cantón of Santa Cruz, Guanacaste, Costa Rica. The experiment was
carried out on a 20m x 20m (400m²) plot, which was divided into 400 basic
experimental units, each corresponding to a one-meter row. The yield of
each unit is recorded in grams.
}
\details{
\itemize{
  \item \strong{Crop:} Corn (Zea mays L.)
  \item \strong{Hybrid:} HS5G
  \item \strong{Planting Density:} 40,000 plants/ha (1m between rows, 0.25m between plants)
  \item \strong{Total Area:} 20m x 20m
  \item \strong{Basic Unit:} 1m x 1m (one linear meter of a row)
  \item \strong{Data:} Yield in grams per basic unit.
  \item \strong{Dimensions:} The matrix has 20 rows (length, Y-axis) and 20 columns (width,
    X-axis).
}
}
\examples{
library("fielddesign")
data("maiz")
# View the first 6 rows and 6 columns.
maiz[1:6, 1:6]

}
\references{
Vargas-Rojas, J. C., & Navarro-Flores, J. R. (2020). Determinación del tamaño y la forma de
unidad experimental, con el método de regresión múltiple, para ensayos de rendimiento de maíz
(Zea mays), Guanacaste, Costa Rica. InterSedes, 21(43), 1-10.
\doi{10.15517/isucr.v21i43.41972}
}
\keyword{datasets}
