% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sums_contour.R
\name{plot_sums_contour}
\alias{plot_sums_contour}
\title{Plot Contour of Sums}
\usage{
plot_sums_contour(sv, width = NULL, length = NULL)
}
\arguments{
\item{sv}{An object of class `sv`, created by `spatial_variation_exhaustive` or
`spatial_variation_tiling` with parameter `return_sums = TRUE`.}

\item{width}{A numeric value specifying the width to filter the `sv` object.
Must correspond to an existing width in `sv$res` and `sv$sums`.}

\item{length}{A numeric value specifying the length to filter the `sv` object.
Must correspond to an existing length in `sv$res` and `sv$sums`}
}
\value{
A `ggplot` object representing the contour plot of sums, or `NULL` invisibly if no data
  matches the provided `width` and `length`, along with a message.
}
\description{
This function generates a contour (tile) plot of sums based on spatial variation data obtained
from `spatial_variation_exhaustive` or `spatial_variation_tiling`.
It visualizes the sum values for a specific `width` and `length` combination.
}
\examples{
\dontrun{
# Generate simulated data.
set.seed(420)
nr <- 10
nc <- 10
x <- matrix(rnorm(nr * nc, 500, 60), nrow = nr, ncol = nc)
# Calculate the exhaustive spatial variation.
sv_exh <- spatial_variation_exhaustive(x, return_sums = TRUE)
# Plot.
plot_sums_contour(sv_exh, 1, 1)
}

}
