% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_yield_contour.R
\name{plot_yield_contour}
\alias{plot_yield_contour}
\title{Plot Yield Contour}
\usage{
plot_yield_contour(x)
}
\arguments{
\item{x}{A numeric matrix representing yield data. Rows are treated as 'Length' and columns as
'Width', with cell values being the 'Value'.}
}
\value{
A `ggplot` object displaying the filled contour plot.
}
\description{
Generates a filled contour plot of yield data from a numeric matrix. The function visualizes
yield values across a 2D grid (Length and Width).
}
\examples{
\dontrun{
# Generate simulated data.
set.seed(420)
nr <- 10
nc <- 10
x <- matrix(rnorm(nr * nc, 500, 60), nrow = nr, ncol = nc)
# Plot.
plot_yield_contour(x)
}

}
