% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_sum.R
\name{sliding_window_sum}
\alias{sliding_window_sum}
\title{Calculate Sliding Window Sums Over a Matrix}
\usage{
sliding_window_sum(x, h, w, combine_orientations = TRUE, return_sums = FALSE)
}
\arguments{
\item{x}{A numeric matrix with all finite numbers.}

\item{h}{The height (number of rows) of the sliding window.}

\item{w}{The width (number of columns) of the sliding window.}

\item{combine_orientations}{A logical value. If `TRUE` (default), the function will also
calculate sums for windows of size `w x h` if `h` and `w` are different, and append them to
the results. If `FALSE`, only `h x w` windows are considered.}

\item{return_sums}{A logical value. If `TRUE`, the function includes a data.frame (`sums`)
detailing the position and size of each window's sum, in addition to the vector of sums
(`res`). If `FALSE` (default), only the vector of sums (`res`) is returned.}
}
\value{
A numeric vector containing the sums of all specified sliding windows.
}
\description{
This function computes the sum of elements within sliding rectangular windows over a numeric
matrix.
It leverages an integral image (cumulative sum matrix) for efficient calculation.
Optionally, it can combine results for both `h x w` and `w x h` window orientations.
}
\keyword{internal}
