% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_numeric_matrix.R
\name{to_numeric_matrix}
\alias{to_numeric_matrix}
\title{Convert to a Numeric Matrix}
\usage{
to_numeric_matrix(x)
}
\arguments{
\item{x}{An R object, typically a matrix or data.frame, whose elements can be coerced to numeric.}
}
\value{
A numeric matrix.
}
\description{
Convert to a Numeric Matrix
}
\keyword{internal}
