% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vigna}
\alias{vigna}
\title{Data of Grain Yield of Cowpea Bean}
\format{
A matrix with 10 rows and 20 columns, where each cell represents the
cowpea bean yield in grams of a 1m² subplot.
}
\source{
Data from an unpublished study. For more details, see the references.
}
\usage{
data("vigna")
}
\description{
This dataset contains yield data from a uniformity trial with cowpea bean
(Vigna unguiculata). The trial was conducted at the Finca Experimental de
Santa Cruz, owned by the University of Costa Rica, located in the cantón
of Santa Cruz, Guanacaste, Costa Rica. The experiment was carried out on a
20m x 10m (200m²) plot, which was divided into 200 basic experimental
units of 1m² each. The yield of each unit is recorded in grams.
}
\details{
\itemize{
  \item \strong{Crop:} Cowpea bean (Vigna unguiculata)
  \item \strong{Location:} Finca Experimental de Santa Cruz, Guanacaste, Costa Rica
  \item \strong{Planting Density:} 100,000 plants/ha
  \item \strong{Total Area:} 20m x 10m
  \item \strong{Basic Unit:} 1m x 1m
  \item \strong{Data:} Yield in grams per basic unit.
  \item \strong{Dimensions:} The matrix has 10 rows (length, Y-axis) and 20 columns (width,
    X-axis).
}
}
\examples{
library("fielddesign")
data("vigna")
# View the first 6 rows and 6 columns.
vigna[1:6, 1:6]

}
\references{
Hernández-González, G., Vargas-Rojas, J. C., & Ríos Gómez, E. (in preparation).
Tamaño y forma de la unidad experimental para ensayos de rendimiento de
frijol caupí (Vigna unguiculata), Santa Cruz, Guanacaste.
}
\keyword{datasets}
